# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for the iserveu1st wallet assessment starter
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  /** @nullable */
  walletId: number | null;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
}

export interface Wallet {
  id: number;
  userId: number;
  userName: string;
  balance: number;
  createdAt: string;
}

export interface WalletInput {
  /** @minimum 1 */
  userId: number;
}

export interface AmountInput {
  /** @exclusiveMinimum 0 */
  amount: number;
}

export interface TransferInput {
  /** @minimum 1 */
  receiverWalletId: number;
  /** @exclusiveMinimum 0 */
  amount: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  DEPOSIT: 'DEPOSIT',
  WITHDRAWAL: 'WITHDRAWAL',
  TRANSFER: 'TRANSFER',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
} as const;

export interface Transaction {
  id: number;
  /** @nullable */
  senderWalletId: number | null;
  /** @nullable */
  receiverWalletId: number | null;
  amount: number;
  type: TransactionType;
  status: TransactionStatus;
  createdAt: string;
  /** @nullable */
  counterpartyName?: string | null;
}

export interface WalletActionResult {
  message: string;
  wallet: Wallet;
  transaction: Transaction;
}

export interface WalletDashboard {
  wallet: Wallet;
  transactions: Transaction[];
  totalDeposited: number;
  totalWithdrawn: number;
  totalTransferred: number;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Conflict
 */
export type ConflictResponse = ErrorResponse;