# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for the iserveu1st wallet assessment starter
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  success: boolean;
  message: string;
}