# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for the iserveu1st wallet assessment starter
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: Date;
  /** @nullable */
  walletId: number | null;
}