# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for the iserveu1st wallet assessment starter
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
}