# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for the iserveu1st wallet assessment starter
 * OpenAPI spec version: 0.1.0
 */

export interface Wallet {
  id: number;
  userId: number;
  userName: string;
  balance: number;
  createdAt: Date;
}