# lib/api-zod/src/generated/types/walletDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for the iserveu1st wallet assessment starter
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';
import type { Wallet } from './wallet';

export interface WalletDashboard {
  wallet: Wallet;
  transactions: Transaction[];
  totalDeposited: number;
  totalWithdrawn: number;
  totalTransferred: number;
}